requirements.txt
fastapi==0.115.5
pydantic==2.10.1
openai==1.57.4
uvicorn==0.27.1
jinja2==3.1.2
python-multipart==0.0.12
python-dotenv==1.0.1
websockets==12.0